requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions==1.23.0
azure-functions-worker==1.23.0
MarkupSafe==3.0.2
Werkzeug==3.1.3
azure-cognitiveservices-vision-computervision==0.9.0
openai

